requirements.txt
# Core dependencies
torch>=1.10.0
transformers>=4.20.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Code parsing
tree-sitter>=0.20.0
tree-sitter-java>=0.20.0

# LLM API
google-generativeai>=0.3.0

# Configuration and validation
pydantic>=1.10.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
hypothesis>=6.50.0

# Utilities
tqdm>=4.60.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For JavaParser fallback
# javalang>=0.13.0
